+++ R/deg_data.R
@@ -1,0 +1,26 @@
+#' @title Differentially expressed genes (DEGs) for volcano_plot function.
+#' @description Differentially expressed genes (DEGs) for volcano_plot function.
+#' @author benben-miao
+#'
+#' @docType data
+#' @format Dataframe: include columns (Genes, Log2FoldChange, Pvalue, Padj), rows (Genes).
+#' @usage data(deg_data)
+#'
+#' @keywords datasets
+#'
+#' @references https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/VolcanoPlot/
+#'
+#' @examples
+#' # 1. Library TOmicsVis package
+#' library(TOmicsVis)
+#'
+#' # 2. Load example data
+#' data(deg_data)
+#'
+#' # 3. View example data
+#' deg_data
+#'
+"deg_data"
+
+# deg_data <- data
+# usethis::use_data(deg_data)