+++ R/deg_data2.R
@@ -1,0 +1,26 @@
+#' @title Differentially expressed genes (DEGs) for ma_plot function.
+#' @description Differentially expressed genes (DEGs) for ma_plot function.
+#' @author benben-miao
+#'
+#' @docType data
+#' @format Dataframe: include columns (Genes, baseMean, Log2FoldChange, Padj), rows (Genes).
+#' @usage data(deg_data2)
+#'
+#' @keywords datasets
+#'
+#' @references https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/MversusA/
+#'
+#' @examples
+#' # 1. Library TOmicsVis package
+#' library(TOmicsVis)
+#'
+#' # 2. Load example data
+#' data(deg_data2)
+#'
+#' # 3. View example data
+#' deg_data2
+#'
+"deg_data2"
+
+# deg_data2 <- data
+# usethis::use_data(deg_data2)